+++ src/blacklist/list.go
@@ -9,7 +9,7 @@ import (
 	"fmt"
 
 	ce "github.com/jeanfrancoisgratton/customError/v3"
-	hftx "github.com/jeanfrancoisgratton/helperFunctions/v4/terminalfx"
+	hftx "github.com/jeanfrancoisgratton/helperFunctions/v5/terminalfx"
 )
 
 // ListAll returns all resources grouped by type.