+++ src/cmd/containers_commands.go
@@ -7,11 +7,12 @@
 package cmd
 
 import (
+	"fmt"
+
 	"dtools2/containers"
 	"dtools2/extras"
 	"dtools2/rest"
-	"dtools2/run"
-	"fmt"
+	"dtools2/run_build"
 
 	"github.com/spf13/cobra"
 )
@@ -288,7 +289,7 @@ var containerAttachCmd = &cobra.Command{
 		}
 		rest.Context = cmd.Context()
 
-		if _, errCode := run.AttachContainer(restClient, args[0]); errCode != nil {
+		if _, errCode := run_build.AttachContainer(restClient, args[0]); errCode != nil {
 			fmt.Println(errCode)
 		}
 		return