+++ src/cmd/env_commands.go
@@ -13,8 +13,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	hf "github.com/jeanfrancoisgratton/helperFunctions/v4"
-	hftx "github.com/jeanfrancoisgratton/helperFunctions/v4/terminalfx"
+	hf "github.com/jeanfrancoisgratton/helperFunctions/v5"
+	hftx "github.com/jeanfrancoisgratton/helperFunctions/v5/terminalfx"
 	"github.com/spf13/cobra"
 )