+++ src/containers/pause_unpause.go
@@ -12,7 +12,7 @@ import (
 	"net/url"
 
 	ce "github.com/jeanfrancoisgratton/customError/v3"
-	hftx "github.com/jeanfrancoisgratton/helperFunctions/v4/terminalfx"
+	hftx "github.com/jeanfrancoisgratton/helperFunctions/v5/terminalfx"
 )
 
 // Pauses one or many containers