+++ src/networks/attach_detach.go
@@ -14,7 +14,7 @@ import (
 	"net/url"
 
 	ce "github.com/jeanfrancoisgratton/customError/v3"
-	hftx "github.com/jeanfrancoisgratton/helperFunctions/v4/terminalfx"
+	hftx "github.com/jeanfrancoisgratton/helperFunctions/v5/terminalfx"
 )
 
 func AttachNetwork(client *rest.Client, network, container string) *ce.CustomError {