+++ src/volumes/prune.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 
 	ce "github.com/jeanfrancoisgratton/customError/v3"
-	hftx "github.com/jeanfrancoisgratton/helperFunctions/v4/terminalfx"
+	hftx "github.com/jeanfrancoisgratton/helperFunctions/v5/terminalfx"
 )
 
 // PruneVolumes removes all unused volumes